Remove dunstrc
This doesn't work with home manager - it looks like the default config
is symlinked from the nix store, and so when HM comes to write this it
throws an error as it is trying to write outside $HOME (following the
symlink). I have moved this config into the home-manager.nix file
instead (stored elsewhere).